+++ gwt-output-qa/src/main/java/arez/gwt/qa/ArezBuildAsserts.java
@@ -198,6 +198,20 @@ public final class ArezBuildAsserts
   }
 
   /**
+   * This assertion verifies that the symbols that are conditional on the `arez.enable_environments`
+   * setting are present if enabled and not present if not enabled.
+   *
+   * @param index   the index that contains all symbols for output target.
+   * @param enabled true if setting is enabled, false otherwise.
+   */
+  public static void assertEnvironmentsOutputs( @Nonnull final SymbolEntryIndex index, final boolean enabled )
+  {
+    index.assertSymbol( "arez\\.ReactionEnvironment", enabled );
+    index.assertSymbol( "arez\\.ArezContext", "_environment", enabled );
+    index.assertSymbol( "arez\\.ArezContext", "_inEnvironmentContext", enabled );
+  }
+
+  /**
    * This assertion verifies that the symbols that are conditional on the `arez.enable_native_components`
    * setting are present if enabled and not present if not enabled.
    *
@@ -249,6 +263,7 @@ public final class ArezBuildAsserts
    *
    * @param index                                the index that contains all symbols for output target.
    * @param areNamesEnabled                      the value of the `arez.enable_names` setting.
+   * @param areEnvironmentsEnabled               the value of the `arez.enable_environments` setting.
    * @param areSpiesEnabled                      the value of the `arez.enable_spies` setting.
    * @param areNativeComponentsEnabled           the value of the `arez.enable_native_components` setting.
    * @param areRegistriesEnabled                 the value of the `arez.enable_registries` setting.
@@ -259,6 +274,7 @@ public final class ArezBuildAsserts
    */
   public static void assertArezOutputs( @Nonnull final SymbolEntryIndex index,
                                         final boolean areNamesEnabled,
+                                        final boolean areEnvironmentsEnabled,
                                         final boolean areSpiesEnabled,
                                         final boolean areNativeComponentsEnabled,
                                         final boolean areRegistriesEnabled,
@@ -276,5 +292,6 @@ public final class ArezBuildAsserts
     assertZoneOutputs( index, areZonesEnabled );
     assertShouldEnforceTransactionTypeOutputs( index, shouldEnforceTransactionType );
     assertCollectionPropertiesUnmodifiableOutputs( index, areCollectionsPropertiesUnmodifiable );
+    assertEnvironmentsOutputs( index, areEnvironmentsEnabled );
   }
 }