+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * **\[core\]** Modify the `ReactionEnvironment` interface so that it executes `Function` and `SafeFunction`
   instances rather than instances of `Procedure`. This will stop `ReactionEnvironment` from being a functional
   interface but is preparation to enable actions to be optionally invoked inside an environment.
+* **\[core\]** Add a compile-time setting `arez.enable_environments` that will mean allows the removal of the
+  `ReactionEnvironment` interface and related code if the feature is not needed.
 
 ### [v0.108](https://github.com/arez/arez/tree/v0.108) (2018-09-27)
 [Full Changelog](https://github.com/arez/arez/compare/v0.107...v0.108)